+++ test/integration/provider-scenarios/android-lifecycle.test.ts
@@ -11,6 +11,7 @@ import {
   assertRpcOk,
 } from './assertions.ts';
 import { createAndroidSettingsWorld, waitForFileContent } from './android-world.ts';
+import { androidAppOwnedSheetXml, androidSystemDialogXml } from './android-dialog-fixtures.ts';
 import { assertAndroidSettingsContract } from './android-settings-contract.ts';
 import { PROVIDER_SCENARIO_ANDROID } from './fixtures.ts';
 import { createProviderScenarioTempPath, withProviderScenarioResource } from './harness.ts';
@@ -238,104 +239,6 @@ test(ANDROID_TOUCH_CONTRACT_EVIDENCE.testName, async () => {
   );
 });
 
-test('Provider-backed integration Android alert handles runtime permission dialog', async () => {
-  const dialog = dismissibleDialog(androidRuntimePermissionXml);
-  await withProviderScenarioResource(
-    async () => await createAndroidSettingsWorld(dialog),
-    async (world) => {
-      const client = world.daemon.client();
-      await client.apps.open({ app: 'com.example.demo', ...world.selection });
-
-      const alertGet = await client.command.alert({ action: 'get', ...world.selection });
-      assert.equal(alertGet.kind, 'alertStatus');
-      assert.deepEqual(alertGet.alert, {
-        title: 'Allow Demo to send you notifications?',
-        buttons: ['Don’t allow', 'Allow'],
-        platform: 'android',
-        source: 'permission',
-        packageName: 'com.google.android.permissioncontroller',
-      });
-
-      const alertAccept = await client.command.alert({ action: 'accept', ...world.selection });
-      assert.equal(alertAccept.kind, 'alertHandled');
-      assert.equal(alertAccept.button, 'Allow');
-      assert.deepEqual(
-        world.adbCalls.filter((call) => call.join(' ') === 'shell input tap 274 638'),
-        [['shell', 'input', 'tap', '274', '638']],
-      );
-
-      dialog.show();
-      const alertDismiss = await client.command.alert({ action: 'dismiss', ...world.selection });
-      assert.equal(alertDismiss.kind, 'alertHandled');
-      assert.equal(alertDismiss.button, 'Don’t allow');
-      assert.deepEqual(
-        world.adbCalls.filter((call) => call.join(' ') === 'shell input tap 116 638'),
-        [['shell', 'input', 'tap', '116', '638']],
-      );
-    },
-  );
-});
-
-test('Provider-backed integration Android alert handles native AlertDialog actions', async () => {
-  const dialog = dismissibleDialog(androidNativeAlertXml);
-  await withProviderScenarioResource(
-    async () => await createAndroidSettingsWorld(dialog),
-    async (world) => {
-      const client = world.daemon.client();
-      await client.apps.open({ app: 'com.example.demo', ...world.selection });
-
-      const alertGet = await client.command.alert({ action: 'get', ...world.selection });
-      assert.deepEqual(alertGet.alert, {
-        title: 'Unsaved changes',
-        message: 'Leave without saving?',
-        buttons: ['Cancel', 'Discard'],
-        platform: 'android',
-        source: 'native-dialog',
-        packageName: 'com.example.demo',
-      });
-
-      const alertAccept = await client.command.alert({ action: 'accept', ...world.selection });
-      assert.equal(alertAccept.button, 'Discard');
-      dialog.show();
-      const alertDismiss = await client.command.alert({ action: 'dismiss', ...world.selection });
-      assert.equal(alertDismiss.button, 'Cancel');
-      assert.deepEqual(
-        world.adbCalls.filter((call) =>
-          ['shell input tap 274 638', 'shell input tap 116 638'].includes(call.join(' ')),
-        ),
-        [
-          ['shell', 'input', 'tap', '274', '638'],
-          ['shell', 'input', 'tap', '116', '638'],
-        ],
-      );
-    },
-  );
-});
-
-test('Provider-backed integration Android alert handles system dialogs', async () => {
-  await withProviderScenarioResource(
-    async () => await createAndroidSettingsWorld(dismissibleDialog(androidSystemDialogXml)),
-    async (world) => {
-      const client = world.daemon.client();
-      await client.apps.open({ app: 'com.example.demo', ...world.selection });
-
-      const alertGet = await client.command.alert({ action: 'get', ...world.selection });
-      assert.deepEqual(alertGet.alert, {
-        title: "Demo isn't responding",
-        message: 'Do you want to close it?',
-        buttons: ['Close app', 'Wait'],
-        platform: 'android',
-        source: 'system-dialog',
-        packageName: 'com.android.systemui',
-      });
-
-      const alertDismiss = await client.command.alert({ action: 'dismiss', ...world.selection });
-      assert.equal(alertDismiss.button, 'Close app');
-      assertCommandCall(world.adbCalls, ['shell', 'input', 'tap', '116', '638']);
-    },
-  );
-});
-
 test('Provider-backed integration Android app-owned ANR recovers before action commands', async () => {
   let anrFocused = true;
   await withProviderScenarioResource(
@@ -406,82 +309,6 @@ test('Provider-backed integration Android external ANR fails with actionable con
         world.adbCalls.filter((call) => call.slice(0, 4).join(' ') === 'shell am start -W').length,
         openCalls,
       );
-    },
-  );
-});
-
-test('Provider-backed integration Android alert dismiss falls back to Back without a dismiss button', async () => {
-  await withProviderScenarioResource(
-    async () => await createAndroidSettingsWorld(dismissibleDialog(androidButtonlessAlertXml)),
-    async (world) => {
-      const client = world.daemon.client();
-      await client.apps.open({ app: 'com.example.demo', ...world.selection });
-
-      const alertDismiss = await client.command.alert({ action: 'dismiss', ...world.selection });
-      assert.equal(alertDismiss.kind, 'alertHandled');
-      assert.equal(alertDismiss.button, 'Back');
-      assertCommandCall(world.adbCalls, ['shell', 'input', 'keyevent', '4']);
-    },
-  );
-});
-
-test('Provider-backed integration Android alert accept fails when the dialog stays visible', async () => {
-  await withProviderScenarioResource(
-    async () => await createAndroidSettingsWorld({ snapshotXml: androidNativeAlertXml }),
-    async (world) => {
-      const client = world.daemon.client();
-      await client.apps.open({ app: 'com.example.demo', ...world.selection });
-
-      await assert.rejects(
-        client.command.alert({ action: 'accept', ...world.selection }),
-        (error: unknown) =>
-          error instanceof Error &&
-          error.message === 'alert accept did not dismiss the visible alert',
-      );
-      assertCommandCall(world.adbCalls, ['shell', 'input', 'tap', '274', '638']);
-    },
-  );
-});
-
-test('Provider-backed integration Android alert wait polls until a dialog appears', async () => {
-  let snapshotCount = 0;
-  await withProviderScenarioResource(
-    async () =>
-      await createAndroidSettingsWorld({
-        snapshotXml: () => {
-          snapshotCount += 1;
-          return snapshotCount === 1 ? androidAppOwnedSheetXml() : androidRuntimePermissionXml();
-        },
-      }),
-    async (world) => {
-      const client = world.daemon.client();
-      await client.apps.open({ app: 'com.example.demo', ...world.selection });
-
-      const alertWait = await client.command.alert({
-        action: 'wait',
-        timeoutMs: 1000,
-        ...world.selection,
-      });
-      assert.equal(alertWait.kind, 'alertWait');
-      // alert now returns the untyped CommandRequestResult bag (its iOS path is a
-      // dynamic runner Record, so the public type is no longer a closed shape).
-      const alertInfo = alertWait.alert as { source?: string } | null | undefined;
-      assert.equal(alertInfo?.source, 'permission');
-      assert.ok(snapshotCount >= 2);
-    },
-  );
-});
-
-test('Provider-backed integration Android alert ignores app-owned sheets', async () => {
-  await withProviderScenarioResource(
-    async () => await createAndroidSettingsWorld({ snapshotXml: androidAppOwnedSheetXml }),
-    async (world) => {
-      const client = world.daemon.client();
-      await client.apps.open({ app: 'com.example.demo', ...world.selection });
-
-      const alertGet = await client.command.alert({ action: 'get', ...world.selection });
-      assert.equal(alertGet.kind, 'alertStatus');
-      assert.equal(alertGet.alert, null);
     },
   );
 });
@@ -667,170 +494,6 @@ async function runAndroidSetupAndInstallWorkflow(
 
   const keyboard = await client.command.keyboard({ action: 'status', ...selection });
   assert.equal(keyboard.visible, false);
-}
-
-function androidRuntimePermissionXml(): string {
-  const packageName = 'com.google.android.permissioncontroller';
-  return androidXml([
-    rootNode(packageName),
-    textNode(
-      1,
-      'Allow Demo to send you notifications?',
-      'com.android.permissioncontroller:id/permission_message',
-      packageName,
-      '[24,300][366,352]',
-    ),
-    buttonNode(
-      2,
-      'Don’t allow',
-      'com.android.permissioncontroller:id/permission_deny_button',
-      '[52,612][180,664]',
-      packageName,
-    ),
-    buttonNode(
-      3,
-      'Allow',
-      'com.android.permissioncontroller:id/permission_allow_button',
-      '[210,612][338,664]',
-      packageName,
-    ),
-    '  </node>',
-  ]);
-}
-
-function androidNativeAlertXml(): string {
-  return androidDialogXml([
-    textNode(2, 'Unsaved changes', 'android:id/alertTitle'),
-    textNode(3, 'Leave without saving?', 'android:id/message'),
-    buttonNode(4, 'Cancel', 'android:id/button2', '[52,612][180,664]'),
-    buttonNode(5, 'Discard', 'android:id/button1', '[210,612][338,664]'),
-  ]);
-}
-
-function androidSystemDialogXml(): string {
-  const packageName = 'com.android.systemui';
-  return androidXml([
-    rootNode(packageName),
-    textNode(1, 'Demo isn&apos;t responding', 'android:id/alertTitle', packageName),
-    textNode(2, 'Do you want to close it?', 'android:id/message', packageName),
-    buttonNode(3, 'Close app', 'android:id/button2', '[52,612][180,664]', packageName),
-    buttonNode(4, 'Wait', 'android:id/button1', '[210,612][338,664]', packageName),
-    '  </node>',
-  ]);
-}
-
-function androidButtonlessAlertXml(): string {
-  return androidDialogXml([
-    textNode(2, 'Unsaved changes', 'android:id/alertTitle'),
-    textNode(3, 'Leave without saving?', 'android:id/message'),
-  ]);
-}
-
-/**
- * A dialog the way a device shows one: in the tree until its button is tapped (or Back is
- * sent), then gone, with the app-owned surface underneath. `show()` brings it back for a
- * second action on the same fixture.
- */
-function dismissibleDialog(dialogXml: () => string) {
-  let visible = true;
-  return {
-    show: () => {
-      visible = true;
-    },
-    snapshotXml: () => (visible ? dialogXml() : androidAppOwnedSheetXml()),
-    onAdbExec: (args: string[]) => {
-      if (args[0] !== 'shell' || args[1] !== 'input') return;
-      if (args[2] === 'tap' || (args[2] === 'keyevent' && args[3] === '4')) visible = false;
-    },
-  };
-}
-
-function androidAppOwnedSheetXml(): string {
-  return androidXml([
-    rootNode('com.example.demo', 'com.example.demo:id/root'),
-    textNode(1, 'Choose an option', 'com.example.demo:id/title'),
-    buttonNode(2, 'Allow', 'com.example.demo:id/allow_button', '[210,612][338,664]'),
-    '  </node>',
-  ]);
-}
-
-function androidDialogXml(children: string[]): string {
-  return androidXml([
-    rootNode(),
-    androidNode({
-      index: 1,
-      id: 'android:id/parentPanel',
-      type: 'android.app.AlertDialog',
-      bounds: '[24,240][366,680]',
-      selfClosing: false,
-    }),
-    ...children,
-    '    </node>',
-    '  </node>',
-  ]);
-}
-
-function androidXml(body: string[]): string {
-  return [
-    '<?xml version="1.0" encoding="UTF-8"?>',
-    '<hierarchy rotation="0">',
-    ...body,
-    '</hierarchy>',
-  ].join('\n');
-}
-
-function rootNode(packageName = 'com.example.demo', id = 'android:id/content'): string {
-  return androidNode({ index: 0, id, type: 'FrameLayout', packageName, selfClosing: false });
-}
-
-function textNode(
-  index: number,
-  text: string,
-  id: string,
-  packageName = 'com.example.demo',
-  bounds?: string,
-): string {
-  return androidNode({ index, text, id, packageName, ...(bounds ? { bounds } : {}) });
-}
-
-function buttonNode(
-  index: number,
-  text: string,
-  id: string,
-  bounds: string,
-  packageName = 'com.example.demo',
-): string {
-  return androidNode({ index, text, id, type: 'Button', packageName, bounds, clickable: true });
-}
-
-function androidNode(options: {
-  index: number;
-  id: string;
-  text?: string;
-  type?: string;
-  packageName?: string;
-  bounds?: string;
-  clickable?: boolean;
-  selfClosing?: boolean;
-}): string {
-  const type = options.type ?? 'TextView';
-  const className = type.includes('.') ? type : `android.widget.${type}`;
-  const tagEnd = options.selfClosing === false ? '>' : ' />';
-  return [
-    `  <node index="${options.index}"`,
-    `text="${options.text ?? ''}"`,
-    `resource-id="${options.id}"`,
-    `class="${className}"`,
-    `package="${options.packageName ?? 'com.example.demo'}"`,
-    'content-desc=""',
-    `bounds="${options.bounds ?? '[48,340][342,392]'}"`,
-    `clickable="${options.clickable ? 'true' : 'false'}"`,
-    'enabled="true"',
-    options.clickable ? 'focusable="true"' : '',
-  ]
-    .filter(Boolean)
-    .join(' ')
-    .concat(tagEnd);
 }
 
 async function runAndroidAppControlAndObservabilityWorkflow(
